htseq-transcriptome v0.1
this commit comes from the successful ramblings on ipython, to
begin creating a successful python script to count reads mapping
to a transcriptome assembly.

Because a transcriptome assembly includes collections of isoforms
from the same gene, any reads that would have mapped uniquely to
a genome assembly - where there are multiple isoforms - would in
a transcriptome assembly map multiply - once to each isoform.

Because these reads are discarded in the classic implementation of
HTSeq, a different implementation needs to be developed to account
for this legitimate side case.