added check-and-skip for empty bundles
during bundle counting, the first bundle in
pair_SAM_alignments( sam_reader, bundle=True ) is always empty,
which was being caught as an error after the assess_bundle function,
since there were no values assigned.

Since it is not an error with the bundle assessment, it seems
appropriate to add a check for empty bundles yielded by
pair_SAM_alignments, so that the "error" does not get output to
screen every time and freak the user out.